Better (correct) code to get Private/Public IP and ignore undefineds.
Work by Costas @ Work
I'm not a Java programmer, so I got a friend at work to help me rewrite
the part that gets the PrivateIpAddress or PublicIpAddress (prefering the
later if it's designed).
BUT also completely ignore any 'undefined' variables (i.e., if they're
BOTH undefined - which they can be if/when the Lambda is run 'to soon',
which they/it is if it's run just before the instance is fully up).